+++ src/main/java/com/example/demo/repository/LoaiSanPhamRepository.java
@@ -3,6 +3,6 @@ package com.example.demo.repository;
 import com.example.demo.entity.LoaiSanPham;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LoaiSanPhamRepository extends JpaRepository<LoaiSanPham, Integer> {
+public interface LoaiSanPhamRepository extends JpaRepository<LoaiSanPham, Long> {
 
 }
